requirements.txt
elementpath~=4.1.0
python-dotenv~=1.0.0
wikibaseintegrator~=0.12.3
